feat: Add "unique" attribute to input fields
This code change adds the "unique" attribute to input fields when the user enters unique values to prevent duplication. The "unique" attribute is set to true when a unique value is entered and false otherwise. This allows developers to style unique inputs differently and improves user experience by providing instant feedback when duplicate values are entered.